requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
speechrecognition>=3.10.0
transformers>=4.30.0
openai-whisper>=20231117
scipy>=1.10.0
matplotlib>=3.7.0
pyaudio>=0.2.11
webrtcvad>=2.0.10
openai>=1.0.0
langchain>=0.0.300
suno-bark>=0.0.1
encodec>=0.1.1
tokenizers>=0.13.0
accelerate>=0.20.0
scikit-learn>=1.2.0